+++ bot/cogs/utility/embeds.py
